+++ pysatMissionPlanning/instruments/__init__.py
@@ -1,8 +1,0 @@
-"""
-pysatMissionPlanning.instruments is a pysat module that provides
-the instrument modules to be used with pysat
-"""
-
-from pysatMissionPlanning.instruments import pysat_ephem, pysat_sgp4
-
-__all__ = ['pysat_ephem', 'pysat_sgp4']