+++ pysatMissions/methods/__init__.py
@@ -1,0 +1,11 @@
+"""
+pysatMissions.methods is a module that provides
+the methods to interface with numerous empirical model packages
+"""
+
+from pysatMissions.methods import aacgmv2
+from pysatMissions.methods import apexpy
+from pysatMissions.methods import pyglow
+from pysatMissions.methods import spacecraft
+
+__all__ = ['aacgmv2', 'apexpy', 'pyglow', 'spacecraft']