+++ pysatMissions/tests/test_methods_spacecraft.py
@@ -3,7 +3,7 @@
 
 import numpy as np
 import pysat
-from pysatMissionPlanning.methods import spacecraft as methsc
+from pysatMissions.methods import spacecraft as mm_sc
 
 
 def add_eci(inst):
@@ -39,7 +39,7 @@ class TestBasics():
 
     def test_calculate_ecef_velocity(self):
         # TODO: check if calculations are correct
-        self.testInst.custom.add(methsc.calculate_ecef_velocity, 'modify')
+        self.testInst.custom.add(mm_sc.calculate_ecef_velocity, 'modify')
         self.testInst.load(date=pysat.datetime(2009, 1, 1))
         targets = ['velocity_ecef_x', 'velocity_ecef_y', 'velocity_ecef_z']
         for target in targets:
@@ -54,8 +54,8 @@ class TestBasics():
 
     def test_add_sc_attitude_vectors(self):
         # TODO: check if calculations are correct
-        self.testInst.custom.add(methsc.calculate_ecef_velocity, 'modify')
-        self.testInst.custom.add(methsc.add_sc_attitude_vectors, 'modify')
+        self.testInst.custom.add(mm_sc.calculate_ecef_velocity, 'modify')
+        self.testInst.custom.add(mm_sc.add_sc_attitude_vectors, 'modify')
         self.testInst.load(date=pysat.datetime(2009, 1, 1))
         targets = ['sc_xhat_ecef_x', 'sc_xhat_ecef_y', 'sc_xhat_ecef_z',
                    'sc_yhat_ecef_x', 'sc_yhat_ecef_y', 'sc_yhat_ecef_z',
@@ -72,10 +72,10 @@ class TestBasics():
 
     def test_project_ecef_vector_onto_sc(self):
         # TODO: check if calculations are correct
-        self.testInst.custom.add(methsc.calculate_ecef_velocity, 'modify')
-        self.testInst.custom.add(methsc.add_sc_attitude_vectors, 'modify')
+        self.testInst.custom.add(mm_sc.calculate_ecef_velocity, 'modify')
+        self.testInst.custom.add(mm_sc.add_sc_attitude_vectors, 'modify')
         self.testInst.custom.add(add_fake_data, 'modify')
-        self.testInst.custom.add(methsc.project_ecef_vector_onto_sc, 'modify',
+        self.testInst.custom.add(mm_sc.project_ecef_vector_onto_sc, 'modify',
                                  'end', 'ax', 'ay', 'az', 'bx', 'by', 'bz')
         self.testInst.load(date=pysat.datetime(2009, 1, 1))
         targets = ['bx', 'by', 'bz']