+++ CHANGELOG.md
@@ -3,6 +3,7 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [0.2.0] - 2019-12-18
+- Renamed as pysatMissions
 - Added method files to access aacgmv2, apexpy, pyglow for any pysat instrument
 - Added method for spacecraft to handle attitude and coordinates
 - Added method for plotting simulated data